refactor: swap-fs, swap-env, swap-serde subcrates, move dependencies into workspace root
