+++ electrum-pool/Cargo.toml
@@ -7,11 +7,11 @@ edition = "2021"
 [dependencies]
 backoff = { version = "0.4", features = ["tokio"] }
 bdk_electrum = { version = "0.19", default-features = false, features = ["use-rustls-ring"] }
-bitcoin = { version = "0.32", features = ["rand", "serde"] }
-futures = { version = "0.3", default-features = false, features = ["std"] }
+bitcoin = { workspace = true }
+futures = { workspace = true }
 once_cell = "1.19"
-tokio = { version = "1", features = ["rt-multi-thread", "time", "macros", "sync"] }
-tracing = { version = "0.1", features = ["attributes"] }
+tokio = { workspace = true }
+tracing = { workspace = true }
 
 [dev-dependencies]
-serde_json = "1"
+serde_json = { workspace = true }
