+++ monero-rpc-pool/Cargo.toml
@@ -9,27 +9,27 @@ name = "monero-rpc-pool"
 path = "src/main.rs"
 
 [dependencies]
-anyhow = "1"
+anyhow = { workspace = true }
 axum = { version = "0.7", features = ["macros"] }
 chrono = { version = "0.4", features = ["serde"] }
 clap = { version = "4.0", features = ["derive"] }
-futures = "0.3"
-monero = { version = "0.12", features = ["serde_support"] }
+futures = { workspace = true }
+monero = { workspace = true }
 monero-rpc = { path = "../monero-rpc" }
-rand = "0.8"
+rand = { workspace = true }
 regex = "1.0"
-reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
+reqwest = { workspace = true, features = ["rustls-tls"] }
+serde = { workspace = true }
+serde_json = { workspace = true }
 sqlx = { version = "0.8", features = ["runtime-tokio-rustls", "sqlite", "chrono", "migrate"] }
-tokio = { version = "1", features = ["full"] }
+tokio = { workspace = true, features = ["full"] }
 tower = "0.4"
 tower-http = { version = "0.5", features = ["cors"] }
-tracing = "0.1"
+tracing = { workspace = true }
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-typeshare = "1.0.3"
+typeshare = { workspace = true }
 url = "2.0"
-uuid = { version = "1.0", features = ["v4"] }
+uuid = { workspace = true }
 
 [dev-dependencies]
 tokio-test = "0.4"
