+++ seed/Cargo.toml
@@ -18,7 +18,7 @@ workspace = true
 [dependencies]
 std-shims = { git = "https://github.com/serai-dex/serai", version = "^0.1.1", default-features = false }
 
-thiserror = { version = "1", default-features = false, optional = true }
+thiserror = { workspace = true, default-features = false, optional = true }
 
 rand_core = { version = "0.6", default-features = false }
 zeroize = { version = "^1.5", default-features = false, features = ["zeroize_derive"] }
