+++ swap-fs/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "swap-fs"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+anyhow = { workspace = true }
+directories-next = "2"
+tracing = { workspace = true }
+
+[lints]
+workspace = true
