+++ swap/Cargo.toml
@@ -12,7 +12,7 @@ name = "swap"
 tauri = ["dep:tauri"]
 
 [dependencies]
-anyhow = "1"
+anyhow = { workspace = true }
 arti-client = { version = "0.25.0", features = ["static-sqlite", "tokio", "rustls", "onion-service-service"], default-features = false }
 async-compression = { version = "0.3", features = ["bzip2", "tokio"] }
 async-trait = "0.1"
@@ -25,7 +25,7 @@ bdk_chain = { version = "0.20" }
 bdk_electrum = { version = "0.19", default-features = false, features = ["use-rustls-ring"] }
 bdk_wallet = { version = "1.0.0-beta.5", features = ["rusqlite", "test-utils"] }
 big-bytes = "1"
-bitcoin = { version = "0.32", features = ["rand", "serde"] }
+bitcoin = { workspace = true }
 bmrng = "0.5.2"
 comfy-table = "7.1"
 config = { version = "0.14", default-features = false, features = ["toml"] }
@@ -34,17 +34,16 @@ curve25519-dalek = { package = "curve25519-dalek-ng", version = "4" }
 data-encoding = "2.6"
 derive_builder = "0.20.2"
 dialoguer = "0.11"
-directories-next = "2"
 ecdsa_fun = { version = "0.10", default-features = false, features = ["libsecp_compat", "serde", "adaptor"] }
 ed25519-dalek = "1"
 electrum-pool = { path = "../electrum-pool" }
 fns = "0.0.7"
-futures = { version = "0.3", default-features = false, features = ["std"] }
-hex = "0.4"
-libp2p = { version = "0.53.2", features = ["tcp", "yamux", "dns", "noise", "request-response", "ping", "rendezvous", "identify", "macros", "cbor", "json", "tokio", "serde", "rsa"] }
+futures = { workspace = true }
+hex = { workspace = true }
+libp2p = { workspace = true, features = ["tcp", "yamux", "dns", "noise", "request-response", "ping", "rendezvous", "identify", "macros", "cbor", "json", "tokio", "serde", "rsa"] }
 libp2p-community-tor = { git = "https://github.com/umgefahren/libp2p-tor", rev = "e6b913e0f1ac1fc90b3ee4dd31b5511140c4a9af", features = ["listen-onion-service"] }
 moka = { version = "0.12", features = ["sync", "future"] }
-monero = { version = "0.12", features = ["serde_support"] }
+monero = { workspace = true }
 monero-rpc = { path = "../monero-rpc" }
 monero-rpc-pool = { path = "../monero-rpc-pool" }
 monero-seed = { version = "0.1.0", path = "../seed" }
@@ -52,42 +51,44 @@ monero-sys = { path = "../monero-sys" }
 once_cell = "1.19"
 pem = "3.0"
 proptest = "1"
-rand = "0.8"
+rand = { workspace = true }
 rand_chacha = "0.3"
 regex = "1.10"
-reqwest = { version = "0.12", features = ["http2", "rustls-tls-native-roots", "stream", "socks"], default-features = false }
+reqwest = { workspace = true, features = ["http2", "rustls-tls-native-roots", "stream", "socks"] }
 rust_decimal = { version = "1", features = ["serde-float"] }
 rust_decimal_macros = "1"
 rustls = { version = "0.23", default-features = false, features = ["ring"] }
 semver = "1.0"
-serde = { version = "1.0", features = ["derive"] }
+serde = { workspace = true }
 serde_cbor = "0.11"
-serde_json = "1"
+serde_json = { workspace = true }
 serde_with = { version = "1", features = ["macros"] }
 sha2 = "0.10"
+swap-serde = { path = "../swap-serde" }
+swap-env = { path = "../swap-env" }
 sigma_fun = { version = "0.7", default-features = false, features = ["ed25519", "serde", "secp256k1", "alloc"] }
 sqlx = { version = "0.8", features = ["sqlite", "runtime-tokio-rustls"] }
 structopt = "0.3"
 strum = { version = "0.26", features = ["derive"] }
 tauri = { version = "2.0", features = ["config-json5"], optional = true, default-features = false }
-thiserror = "1"
+thiserror = { workspace = true }
 time = "0.3"
-tokio = { version = "1", features = ["rt-multi-thread", "time", "macros", "sync", "process", "fs", "net", "parking_lot", "rt"] }
+tokio = { workspace = true, features = ["process", "fs", "net", "parking_lot", "rt"] }
 tokio-tungstenite = { version = "0.15", features = ["rustls-tls"] }
 tokio-util = { version = "0.7", features = ["io", "codec", "rt"] }
-toml = "0.8"
 tor-rtcompat = { version = "0.25.0", features = ["tokio"] }
 tower = { version = "0.4.13", features = ["full"] }
 tower-http = { version = "0.3.4", features = ["full"] }
-tracing = { version = "0.1", features = ["attributes"] }
+tracing = { workspace = true }
 tracing-appender = "0.2"
 tracing-subscriber = { version = "0.3", default-features = false, features = ["fmt", "ansi", "env-filter", "time", "tracing-log", "json"] }
-typeshare = "1.0.4"
+typeshare = { workspace = true }
 unsigned-varint = { version = "0.8.0", features = ["codec", "asynchronous_codec"] }
-url = { version = "2", features = ["serde"] }
-uuid = { version = "1.9", features = ["serde", "v4"] }
+url = { workspace = true }
+uuid = { workspace = true, features = ["serde"] }
 void = "1"
 zeroize = "1.8.1"
+swap-fs = { path = "../swap-fs" }
 
 [target.'cfg(not(windows))'.dependencies]
 tokio-tar = "0.3"
@@ -107,5 +108,5 @@ tempfile = "3"
 testcontainers = "0.15"
 
 [build-dependencies]
-anyhow = "1"
+anyhow = { workspace = true }
 vergen = { version = "8.3", default-features = false, features = ["build", "git", "git2"] }
