+++ swap/src/asb/event_loop.rs
@@ -7,7 +7,8 @@ use crate::network::transfer_proof;
 use crate::protocol::alice::swap::has_already_processed_enc_sig;
 use crate::protocol::alice::{AliceState, ReservesMonero, State3, Swap};
 use crate::protocol::{Database, State};
-use crate::{bitcoin, env, kraken, monero};
+use crate::{bitcoin, kraken, monero};
+use swap_env::env;
 use anyhow::{anyhow, Context, Result};
 use futures::future;
 use futures::future::{BoxFuture, FutureExt};
