+++ swap/src/lib.rs
@@ -21,8 +21,6 @@ pub mod bitcoin;
 pub mod cli;
 pub mod common;
 pub mod database;
-pub mod env;
-pub mod fs;
 pub mod kraken;
 pub mod libp2p_ext;
 pub mod monero;
