+++ swap/src/network/swap_setup/alice.rs
@@ -5,7 +5,8 @@ use crate::network::swap_setup::{
 };
 use crate::protocol::alice::{State0, State3};
 use crate::protocol::{Message0, Message2, Message4};
-use crate::{asb, bitcoin, env, monero};
+use crate::{asb, bitcoin, monero};
+use swap_env::env;
 use anyhow::{anyhow, Context, Result};
 use futures::future::{BoxFuture, OptionFuture};
 use futures::AsyncWriteExt;
