+++ swap/src/network/swarm.rs
@@ -2,7 +2,8 @@ use crate::asb::{LatestRate, RendezvousNode};
 use crate::libp2p_ext::MultiAddrExt;
 use crate::network::rendezvous::XmrBtcNamespace;
 use crate::seed::Seed;
-use crate::{asb, bitcoin, cli, env};
+use crate::{asb, bitcoin, cli};
+use swap_env::env;
 use anyhow::Result;
 use arti_client::TorClient;
 use libp2p::swarm::NetworkBehaviour;
