+++ swap/src/protocol/bob/state.rs
@@ -1,4 +1,4 @@
-use crate::bitcoin::address_serde;
+use swap_serde::bitcoin::address_serde;
 use crate::bitcoin::wallet::{EstimateFeeRate, Subscription};
 use crate::bitcoin::{
     self, current_epoch, CancelTimelock, ExpiredTimelocks, PunishTimelock, Transaction, TxCancel,
@@ -6,7 +6,7 @@ use crate::bitcoin::{
 };
 use crate::monero::wallet::WatchRequest;
 use crate::monero::{self, MoneroAddressPool, TxHash};
-use crate::monero::{monero_private_key, TransferProof};
+use crate::monero::{TransferProof};
 use crate::monero_ext::ScalarExt;
 use crate::protocol::{Message0, Message1, Message2, Message3, Message4, CROSS_CURVE_PROOF_SYSTEM};
 use anyhow::{anyhow, bail, Context, Result};
@@ -685,7 +685,7 @@ impl State4 {
 
 #[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
 pub struct State5 {
-    #[serde(with = "monero_private_key")]
+    #[serde(with = "swap_serde::monero::private_key")]
     s_a: monero::PrivateKey,
     s_b: monero::Scalar,
     v: monero::PrivateViewKey,
