+++ swap/tests/harness/mod.rs
@@ -21,8 +21,8 @@ use swap::asb::FixedRate;
 use swap::bitcoin::{CancelTimelock, PunishTimelock};
 use swap::cli::api;
 use swap::database::{AccessMode, SqliteDatabase};
-use swap::env::{Config, GetConfig};
-use swap::fs::ensure_directory_exists;
+use swap_env::env::{Config, GetConfig};
+use swap_fs::ensure_directory_exists;
 use swap::monero::wallet::no_listener;
 use swap::monero::Wallets;
 use swap::network::rendezvous::XmrBtcNamespace;
@@ -31,7 +31,8 @@ use swap::protocol::alice::{AliceState, Swap};
 use swap::protocol::bob::BobState;
 use swap::protocol::{alice, bob, Database};
 use swap::seed::Seed;
-use swap::{asb, bitcoin, cli, env, monero};
+use swap::{asb, bitcoin, cli, monero};
+use swap_env::env;
 use tempfile::{NamedTempFile, TempDir};
 use testcontainers::clients::Cli;
 use testcontainers::{Container, RunnableImage};
